requirements.txt
streamlit
qrcode
Pillow
opencv-python-headless
numpy
pandas




